+++ SMS/SMS.Business/IInvoiceDetailManagement.cs
@@ -5,7 +5,7 @@ namespace SMS.Business
 {
     public interface IInvoiceDetailManagement : IBaseManagement<InvoiceDetailDto, long>
     {
-        InvoiceDetailDto AddProductToInvoiceTable(long invoiceTableID, long productID, int quantity);
+        InvoiceDetailDto AddProductToInvoiceTable(long invoiceTableID, long productID, decimal quantity);
         InvoiceDetailDto UpdateProductToInvoiceTable(long invoiceDetailID, string columnName, string value);
     }
 }
